ICMSLST-522 Add sanction applicant details view unit tests
Also

* Refactor of firearms tests
